+++ src/main/java/com/quarkbau/monolith/planning/dto/EmployeeDTO.java
@@ -22,5 +22,5 @@ public class EmployeeDTO {
     private CompanyRole role;
     private Boolean enabled;
     private Long organizationId;
-    private BigInteger subcontractorId;
+    private Long subcontractorId;
 }
